+++ lib/let_me/literal.ex
@@ -1,9 +1,0 @@
-defmodule LetMe.Literal do
-  @moduledoc """
-  Struct that represents an authorization rule that evaluates to a fixed value.
-  """
-
-  @type t :: %__MODULE__{passed?: boolean}
-
-  defstruct [:passed?]
-end