+++ test/let_me/policy_test.exs
@@ -4,16 +4,16 @@ defmodule LetMe.PolicyTest do
 
   import ExUnit.CaptureLog
 
-  alias LetMe.AllOf
-  alias LetMe.AnyOf
-  alias LetMe.Check
-  alias LetMe.Literal
-  alias LetMe.Not
   alias LetMe.Rule
   alias LetMe.UnauthorizedError
   alias MyApp.Blog.Article
   alias MyApp.Policy
   alias MyApp.TestPolicy
+  alias Spek.AllOf
+  alias Spek.AnyOf
+  alias Spek.Check
+  alias Spek.Literal
+  alias Spek.Not
 
   defmodule TestPolicy do
     use LetMe.Policy, check_module: MyApp.Checks, error: :detailed
@@ -294,8 +294,16 @@ defmodule LetMe.PolicyTest do
                    action: :create,
                    expression: %AnyOf{
                      children: [
-                       %Check{name: :role, arg: :admin},
-                       %Check{name: :role, arg: :writer}
+                       %Check{
+                         module: MyApp.Checks,
+                         fun: :role,
+                         args: [{:ctx, :subject}, {:ctx, :object}, :admin]
+                       },
+                       %Check{
+                         module: MyApp.Checks,
+                         fun: :role,
+                         args: [{:ctx, :subject}, {:ctx, :object}, :writer]
+                       }
                      ]
                    },
                    name: :article_create,
@@ -304,14 +312,18 @@ defmodule LetMe.PolicyTest do
                  },
                  %Rule{
                    action: :update,
-                   expression: %Check{name: :own_resource},
+                   expression: %Check{
+                     module: MyApp.Checks,
+                     fun: :own_resource,
+                     args: [{:ctx, :subject}, {:ctx, :object}]
+                   },
                    name: :article_update,
                    object: :article,
                    pre_hooks: [:preload_groups]
                  },
                  %Rule{
                    action: :view,
-                   expression: %Literal{passed?: true},
+                   expression: %Literal{result: true, satisfied?: true},
                    description:
                      "allows to view an article and the list of articles",
                    name: :article_view,
@@ -322,8 +334,18 @@ defmodule LetMe.PolicyTest do
                    action: :delete,
                    expression: %AllOf{
                      children: [
-                       %Not{expression: %Check{name: :same_user}},
-                       %Check{name: :role, arg: :admin}
+                       %Not{
+                         expression: %Check{
+                           module: MyApp.Checks,
+                           fun: :same_user,
+                           args: [{:ctx, :subject}, {:ctx, :object}]
+                         }
+                       },
+                       %Check{
+                         module: MyApp.Checks,
+                         fun: :role,
+                         args: [{:ctx, :subject}, {:ctx, :object}, :admin]
+                       }
                      ]
                    },
                    name: :user_delete,
@@ -338,8 +360,16 @@ defmodule LetMe.PolicyTest do
                    action: :list,
                    expression: %AnyOf{
                      children: [
-                       %Check{name: :role, arg: :admin},
-                       %Check{name: :role, arg: :client}
+                       %Check{
+                         module: MyApp.Checks,
+                         fun: :role,
+                         args: [{:ctx, :subject}, {:ctx, :object}, :admin]
+                       },
+                       %Check{
+                         module: MyApp.Checks,
+                         fun: :role,
+                         args: [{:ctx, :subject}, {:ctx, :object}, :client]
+                       }
                      ]
                    },
                    name: :user_list,
@@ -348,7 +378,11 @@ defmodule LetMe.PolicyTest do
                  },
                  %Rule{
                    action: :remove,
-                   expression: %Check{name: :role, arg: :super_admin},
+                   expression: %Check{
+                     module: MyApp.Checks,
+                     fun: :role,
+                     args: [{:ctx, :subject}, {:ctx, :object}, :super_admin]
+                   },
                    name: :user_remove,
                    object: :user,
                    pre_hooks: [],
@@ -358,14 +392,30 @@ defmodule LetMe.PolicyTest do
                    action: :view,
                    expression: %AnyOf{
                      children: [
-                       %Check{name: :role, arg: :admin},
+                       %Check{
+                         module: MyApp.Checks,
+                         fun: :role,
+                         args: [{:ctx, :subject}, {:ctx, :object}, :admin]
+                       },
                        %AllOf{
                          children: [
-                           %Check{name: :role, arg: :client},
-                           %Check{name: :same_company}
+                           %Check{
+                             module: MyApp.Checks,
+                             fun: :role,
+                             args: [{:ctx, :subject}, {:ctx, :object}, :client]
+                           },
+                           %Check{
+                             module: MyApp.Checks,
+                             fun: :same_company,
+                             args: [{:ctx, :subject}, {:ctx, :object}]
+                           }
                          ]
                        },
-                       %Check{name: :same_user}
+                       %Check{
+                         module: MyApp.Checks,
+                         fun: :same_user,
+                         args: [{:ctx, :subject}, {:ctx, :object}]
+                       }
                      ]
                    },
                    name: :user_view,
@@ -412,7 +462,7 @@ defmodule LetMe.PolicyTest do
              ] =
                Policy.list_rules(
                  check: fn
-                   %Check{name: :same_user} -> true
+                   %Check{module: MyApp.Checks, fun: :same_user} -> true
                    %Check{} -> false
                  end
                )
@@ -489,10 +539,18 @@ defmodule LetMe.PolicyTest do
     test "returns rule" do
       assert Policy.get_rule(:article_create) == %Rule{
                action: :create,
-               expression: %LetMe.AnyOf{
+               expression: %Spek.AnyOf{
                  children: [
-                   %Check{name: :role, arg: :admin},
-                   %Check{name: :role, arg: :writer}
+                   %Check{
+                     module: MyApp.Checks,
+                     fun: :role,
+                     args: [{:ctx, :subject}, {:ctx, :object}, :admin]
+                   },
+                   %Check{
+                     module: MyApp.Checks,
+                     fun: :role,
+                     args: [{:ctx, :subject}, {:ctx, :object}, :writer]
+                   }
                  ]
                },
                name: :article_create,
@@ -514,10 +572,18 @@ defmodule LetMe.PolicyTest do
                   action: :create,
                   expression: %AnyOf{
                     children: [
-                      %Check{name: :role, arg: :admin},
-                      %Check{name: :role, arg: :writer}
+                      %Check{
+                        module: MyApp.Checks,
+                        fun: :role,
+                        args: [{:ctx, :subject}, {:ctx, :object}, :admin]
+                      },
+                      %Check{
+                        module: MyApp.Checks,
+                        fun: :role,
+                        args: [{:ctx, :subject}, {:ctx, :object}, :writer]
+                      }
                     ],
-                    passed?: nil
+                    satisfied?: nil
                   },
                   name: :article_create,
                   object: :article,
@@ -536,8 +602,16 @@ defmodule LetMe.PolicyTest do
                action: :create,
                expression: %AnyOf{
                  children: [
-                   %Check{name: :role, arg: :admin},
-                   %Check{name: :role, arg: :writer}
+                   %Check{
+                     module: MyApp.Checks,
+                     fun: :role,
+                     args: [{:ctx, :subject}, {:ctx, :object}, :admin]
+                   },
+                   %Check{
+                     module: MyApp.Checks,
+                     fun: :role,
+                     args: [{:ctx, :subject}, {:ctx, :object}, :writer]
+                   }
                  ]
                },
                name: :article_create,
@@ -567,9 +641,10 @@ defmodule LetMe.PolicyTest do
                %{user_id: 2}
              ) == %UnauthorizedError{
                expression: %Check{
-                 arg: nil,
-                 name: :own_resource,
-                 passed?: false,
+                 module: MyApp.Checks,
+                 fun: :own_resource,
+                 args: [{:ctx, :subject}, {:ctx, :object}],
+                 satisfied?: false,
                  result: false
                },
                message: "unauthorized"
@@ -593,10 +668,11 @@ defmodule LetMe.PolicyTest do
              ) == %UnauthorizedError{
                message: "unauthorized",
                expression: %Check{
-                 name: :role,
-                 arg: :editor,
+                 module: MyApp.Checks,
+                 fun: :role,
+                 args: [{:ctx, :subject}, {:ctx, :object}, :editor],
                  result: false,
-                 passed?: false
+                 satisfied?: false
                }
              }
     end
@@ -608,7 +684,7 @@ defmodule LetMe.PolicyTest do
       assert unauthorized_error(TestPolicy, :simple_allow_false, %{}) ==
                %UnauthorizedError{
                  message: "unauthorized",
-                 expression: %Literal{passed?: false}
+                 expression: %Literal{result: false, satisfied?: false}
                }
 
       assert unauthorized_error(
@@ -617,7 +693,7 @@ defmodule LetMe.PolicyTest do
                %{}
              ) == %UnauthorizedError{
                message: "unauthorized",
-               expression: %Literal{passed?: false}
+               expression: %Literal{result: false, satisfied?: false}
              }
     end
 
@@ -633,9 +709,10 @@ defmodule LetMe.PolicyTest do
                %{role: :writer}
              ) == %UnauthorizedError{
                expression: %Check{
-                 arg: :admin,
-                 name: :role,
-                 passed?: false,
+                 module: MyApp.Checks,
+                 fun: :role,
+                 args: [{:ctx, :subject}, {:ctx, :object}, :admin],
+                 satisfied?: false,
                  result: false
                },
                message: "unauthorized"
@@ -649,7 +726,7 @@ defmodule LetMe.PolicyTest do
                :simple_allow_false_combined,
                %{role: :admin}
              ) == %UnauthorizedError{
-               expression: %Literal{passed?: false},
+               expression: %Literal{result: false, satisfied?: false},
                message: "unauthorized"
              }
     end
@@ -664,12 +741,13 @@ defmodule LetMe.PolicyTest do
                message: "unauthorized",
                expression: %Not{
                  expression: %Check{
-                   name: :same_user,
-                   arg: nil,
+                   module: MyApp.Checks,
+                   fun: :same_user,
+                   args: [{:ctx, :subject}, {:ctx, :object}],
                    result: true,
-                   passed?: true
+                   satisfied?: true
                  },
-                 passed?: false
+                 satisfied?: false
                }
              }
 
@@ -692,12 +770,13 @@ defmodule LetMe.PolicyTest do
                message: "unauthorized",
                expression: %Not{
                  expression: %Check{
-                   name: :role,
-                   arg: :writer,
-                   passed?: true,
+                   module: MyApp.Checks,
+                   fun: :role,
+                   args: [{:ctx, :subject}, {:ctx, :object}, :writer],
+                   satisfied?: true,
                    result: true
                  },
-                 passed?: false
+                 satisfied?: false
                }
              }
     end
@@ -706,7 +785,7 @@ defmodule LetMe.PolicyTest do
       assert unauthorized_error(TestPolicy, :simple_deny_true, %{}) ==
                %UnauthorizedError{
                  message: "unauthorized",
-                 expression: %Literal{passed?: false}
+                 expression: %Literal{result: false, satisfied?: false}
                }
 
       assert_authorized TestPolicy, :simple_deny_false, %{}
@@ -720,7 +799,7 @@ defmodule LetMe.PolicyTest do
                %{id: 1}
              ) == %UnauthorizedError{
                message: "unauthorized",
-               expression: %Literal{passed?: false}
+               expression: %Literal{result: false, satisfied?: false}
              }
 
       assert unauthorized_error(
@@ -730,7 +809,7 @@ defmodule LetMe.PolicyTest do
                %{id: 2}
              ) == %UnauthorizedError{
                message: "unauthorized",
-               expression: %Literal{passed?: false}
+               expression: %Literal{result: false, satisfied?: false}
              }
     end
 
@@ -738,13 +817,13 @@ defmodule LetMe.PolicyTest do
       assert unauthorized_error(TestPolicy, :simple_no_checks, %{}) ==
                %UnauthorizedError{
                  message: "unauthorized",
-                 expression: %Literal{passed?: false}
+                 expression: %Literal{result: false, satisfied?: false}
                }
 
       assert unauthorized_error(TestPolicy, :simple_empty_list_check, %{}) ==
                %UnauthorizedError{
                  message: "unauthorized",
-                 expression: %Literal{passed?: false}
+                 expression: %Literal{result: false, satisfied?: false}
                }
     end
 
@@ -760,10 +839,11 @@ defmodule LetMe.PolicyTest do
              ) == %UnauthorizedError{
                message: "unauthorized",
                expression: %Check{
-                 name: :role,
-                 arg: :admin,
+                 module: MyApp.Checks,
+                 fun: :role,
+                 args: [{:ctx, :subject}, {:ctx, :object}, :admin],
                  result: false,
-                 passed?: false
+                 satisfied?: false
                }
              }
     end
@@ -786,8 +866,10 @@ defmodule LetMe.PolicyTest do
                %{user_id: 2}
              ) == %UnauthorizedError{
                expression: %Check{
-                 name: :own_resource,
-                 passed?: false,
+                 module: MyApp.Checks,
+                 fun: :own_resource,
+                 args: [{:ctx, :subject}, {:ctx, :object}],
+                 satisfied?: false,
                  result: false
                },
                message: "unauthorized"
@@ -800,9 +882,10 @@ defmodule LetMe.PolicyTest do
                %{user_id: 2}
              ) == %UnauthorizedError{
                expression: %Check{
-                 name: :own_resource,
-                 arg: nil,
-                 passed?: false,
+                 module: MyApp.Checks,
+                 fun: :own_resource,
+                 args: [{:ctx, :subject}, {:ctx, :object}],
+                 satisfied?: false,
                  result: false
                },
                message: "unauthorized"
@@ -819,7 +902,10 @@ defmodule LetMe.PolicyTest do
                assert TestPolicy.authorize(:does_not_exist, %{}) ==
                         {:error,
                          %UnauthorizedError{
-                           expression: %LetMe.Literal{passed?: false},
+                           expression: %Spek.Literal{
+                             result: {:error, :unknown_policy},
+                             satisfied?: false
+                           },
                            message: "unauthorized"
                          }}
              end) =~ "Permission checked for rule that does not exist"
@@ -839,12 +925,14 @@ defmodule LetMe.PolicyTest do
                %{user_id: 2}
              ) == %UnauthorizedError{
                expression: %AllOf{
-                 passed?: false,
+                 satisfied?: false,
                  children: [
-                   %LetMe.Check{
-                     name: :own_resource,
+                   %Spek.Check{
+                     module: MyApp.Checks,
+                     fun: :own_resource,
+                     args: [{:ctx, :subject}, {:ctx, :object}],
                      result: false,
-                     passed?: false
+                     satisfied?: false
                    }
                  ]
                },
@@ -860,19 +948,21 @@ defmodule LetMe.PolicyTest do
                expression: %AllOf{
                  children: [
                    %Check{
-                     name: :own_resource,
-                     arg: nil,
+                     module: MyApp.Checks,
+                     fun: :own_resource,
+                     args: [{:ctx, :subject}, {:ctx, :object}],
                      result: true,
-                     passed?: true
+                     satisfied?: true
                    },
                    %Check{
-                     name: :role,
-                     arg: :editor,
+                     module: MyApp.Checks,
+                     fun: :role,
+                     args: [{:ctx, :subject}, {:ctx, :object}, :editor],
                      result: false,
-                     passed?: false
+                     satisfied?: false
                    }
                  ],
-                 passed?: false
+                 satisfied?: false
                },
                message: "unauthorized"
              }
@@ -883,14 +973,15 @@ defmodule LetMe.PolicyTest do
                %{id: 1, role: :writer},
                %{user_id: 2}
              ) == %UnauthorizedError{
-               expression: %LetMe.AllOf{
-                 passed?: false,
+               expression: %Spek.AllOf{
+                 satisfied?: false,
                  children: [
-                   %LetMe.Check{
-                     name: :own_resource,
-                     arg: nil,
+                   %Spek.Check{
+                     module: MyApp.Checks,
+                     fun: :own_resource,
+                     args: [{:ctx, :subject}, {:ctx, :object}],
                      result: false,
-                     passed?: false
+                     satisfied?: false
                    }
                  ]
                },
@@ -925,19 +1016,21 @@ defmodule LetMe.PolicyTest do
                expression: %AnyOf{
                  children: [
                    %Check{
-                     name: :role,
-                     arg: :editor,
+                     module: MyApp.Checks,
+                     fun: :role,
+                     args: [{:ctx, :subject}, {:ctx, :object}, :editor],
                      result: false,
-                     passed?: false
+                     satisfied?: false
                    },
                    %Check{
-                     name: :own_resource,
-                     arg: nil,
+                     module: MyApp.Checks,
+                     fun: :own_resource,
+                     args: [{:ctx, :subject}, {:ctx, :object}],
                      result: false,
-                     passed?: false
+                     satisfied?: false
                    }
                  ],
-                 passed?: false
+                 satisfied?: false
                },
                message: "unauthorized"
              }
@@ -956,24 +1049,26 @@ defmodule LetMe.PolicyTest do
                  children: [
                    %Not{
                      expression: %Check{
-                       name: :same_user,
-                       arg: nil,
+                       module: MyApp.Checks,
+                       fun: :same_user,
+                       args: [{:ctx, :subject}, {:ctx, :object}],
                        result: true,
-                       passed?: true
+                       satisfied?: true
                      },
-                     passed?: false
+                     satisfied?: false
                    },
                    %Not{
                      expression: %Check{
-                       name: :role,
-                       arg: :writer,
+                       module: MyApp.Checks,
+                       fun: :role,
+                       args: [{:ctx, :subject}, {:ctx, :object}, :writer],
                        result: true,
-                       passed?: true
+                       satisfied?: true
                      },
-                     passed?: false
+                     satisfied?: false
                    }
                  ],
-                 passed?: false
+                 satisfied?: false
                },
                message: "unauthorized"
              }
@@ -1007,14 +1102,16 @@ defmodule LetMe.PolicyTest do
                  children: [
                    %Not{
                      expression: %Check{
-                       name: :same_user,
-                       passed?: true,
+                       module: MyApp.Checks,
+                       fun: :same_user,
+                       args: [{:ctx, :subject}, {:ctx, :object}],
+                       satisfied?: true,
                        result: true
                      },
-                     passed?: false
+                     satisfied?: false
                    }
                  ],
-                 passed?: false
+                 satisfied?: false
                },
                message: "unauthorized"
              }
@@ -1031,15 +1128,17 @@ defmodule LetMe.PolicyTest do
                %{id: 1}
              ) == %UnauthorizedError{
                expression: %AllOf{
-                 passed?: false,
+                 satisfied?: false,
                  children: [
                    %Not{
                      expression: %Check{
-                       name: :same_user,
+                       module: MyApp.Checks,
+                       fun: :same_user,
+                       args: [{:ctx, :subject}, {:ctx, :object}],
                        result: true,
-                       passed?: true
+                       satisfied?: true
                      },
-                     passed?: false
+                     satisfied?: false
                    }
                  ]
                },
@@ -1053,25 +1152,27 @@ defmodule LetMe.PolicyTest do
                %{id: 2}
              ) == %UnauthorizedError{
                expression: %AllOf{
-                 passed?: false,
+                 satisfied?: false,
                  children: [
-                   %LetMe.Not{
+                   %Spek.Not{
                      expression: %Check{
-                       name: :same_user,
-                       arg: nil,
+                       module: MyApp.Checks,
+                       fun: :same_user,
+                       args: [{:ctx, :subject}, {:ctx, :object}],
                        result: false,
-                       passed?: false
+                       satisfied?: false
                      },
-                     passed?: true
+                     satisfied?: true
                    },
-                   %LetMe.Not{
+                   %Spek.Not{
                      expression: %Check{
-                       name: :role,
-                       arg: :writer,
+                       module: MyApp.Checks,
+                       fun: :role,
+                       args: [{:ctx, :subject}, {:ctx, :object}, :writer],
                        result: true,
-                       passed?: true
+                       satisfied?: true
                      },
-                     passed?: false
+                     satisfied?: false
                    }
                  ]
                },
@@ -1210,15 +1311,17 @@ defmodule LetMe.PolicyTest do
                %{role: :admin, state: :suspended}
              ) == %UnauthorizedError{
                expression: %AllOf{
-                 passed?: false,
+                 satisfied?: false,
                  children: [
-                   %LetMe.Not{
-                     expression: %LetMe.Check{
-                       name: :user_suspended,
+                   %Spek.Not{
+                     expression: %Spek.Check{
+                       module: MyApp.Checks,
+                       fun: :user_suspended,
+                       args: [{:ctx, :subject}, {:ctx, :object}],
                        result: {:ok, "user suspended"},
-                       passed?: true
+                       satisfied?: true
                      },
-                     passed?: false
+                     satisfied?: false
                    }
                  ]
                },
@@ -1234,20 +1337,23 @@ defmodule LetMe.PolicyTest do
                  children: [
                    %Not{
                      expression: %Check{
-                       name: :user_suspended,
+                       module: MyApp.Checks,
+                       fun: :user_suspended,
+                       args: [{:ctx, :subject}, {:ctx, :object}],
                        result: :error,
-                       passed?: false
+                       satisfied?: false
                      },
-                     passed?: true
+                     satisfied?: true
                    },
                    %Check{
-                     arg: :admin,
-                     name: :role_with_reason,
+                     module: MyApp.Checks,
+                     fun: :role_with_reason,
+                     args: [{:ctx, :subject}, {:ctx, :object}, :admin],
                      result: {:error, "writer not allowed"},
-                     passed?: false
+                     satisfied?: false
                    }
                  ],
-                 passed?: false
+                 satisfied?: false
                },
                message: "unauthorized"
              }
@@ -1255,7 +1361,7 @@ defmodule LetMe.PolicyTest do
 
     test "can customize error response" do
       # TestPolicy uses detailed errors (error: :detailed)
-      assert {:error, %UnauthorizedError{expression: %LetMe.Literal{}}} =
+      assert {:error, %UnauthorizedError{expression: %Spek.Literal{}}} =
                TestPolicy.authorize(:simple_allow_false, %{})
 
       # PolicyShort uses default error
@@ -1269,7 +1375,7 @@ defmodule LetMe.PolicyTest do
                  error: :unauthorized
                )
 
-      assert {:error, %UnauthorizedError{expression: %LetMe.AnyOf{}}} =
+      assert {:error, %UnauthorizedError{expression: %Spek.AnyOf{}}} =
                MyApp.PolicyShort.authorize(:article_create, %{}, %{},
                  error: :detailed
                )
@@ -1300,7 +1406,7 @@ defmodule LetMe.PolicyTest do
 
     test "can opt-in to / out of detailed errors" do
       # TestPolicy uses detailed errors (error: :detailed)
-      assert %LetMe.UnauthorizedError{expression: %LetMe.Literal{}} =
+      assert %LetMe.UnauthorizedError{expression: %Spek.Literal{}} =
                assert_raise(
                  LetMe.UnauthorizedError,
                  "unauthorized",
@@ -1332,7 +1438,7 @@ defmodule LetMe.PolicyTest do
                  end
                )
 
-      assert %LetMe.UnauthorizedError{expression: %LetMe.AnyOf{}} =
+      assert %LetMe.UnauthorizedError{expression: %Spek.AnyOf{}} =
                assert_raise(
                  LetMe.UnauthorizedError,
                  "unauthorized",